Chaos: HashTable implementation for AABBTree payload tracking to reduce memory costs of AABB trees in when used as a local sub-acceleration.
Users of TAABBTree can provide a storage policy that currently only handles storage requirements for the PayloadToInfo mapping store. The main scene
uses a flat array sized to the unique indices of the particles inside it. This is OK for the main scene but when used as a sub-acceleration later in a run
when the unique index counter is very high there's a lot of empty space in the payload map.

Here cluster unions defer to a fast hash table for this mapping rather than a flat array. Because cluster unions are usually small, at least smaller than
the whole scene the lookup in the map shouldn't often chain and should be relatively fast.

#rb Cedric.Caillaud, Chris.Caulfield, Jaco.VanDyk, Michael.Bao

[CL 27592731 by benn gallagher in ue5-main branch]