TEDS can now delay adding/removing columns if needed.
TEDS will now check if a row has been initialized or not. If the row has been created but not initialized it will not delay adding/removing columns until after the row has been initialized
This functionality is currently limited to adding columns without data. Retrieving columns in this state is still not possible.
This is an initial step in a larger future effort to have a uniform handling of commands throughout TEDS' direct calls and query callbacks.

#jira UE-193544
#rb brooke.hubert

[CL 27722924 by ronald koppers in ue5-main branch]