Fixed the same object getting (un)registered with TEDS Compat causing a crash.
In some cases the UI can rapidly cause objects to be created  and destroyed, for instance by dragging sliders around. It's possible for this to happen within the same frame, causing the same object to be unregistered and then registered again multiple times. This in turn causes the same object to be pending registration multiple times. The first registration will succeed but not the following ones as there's already an object registered. This change removes the unregistered version from the queue before processing.

#jira UE-194166, UE-193907
#rb logan.buchy brooke.hubert

[CL 27763660 by ronald koppers in ue5-main branch]