Critical fix to how Mass handles mutating command-instigated operations on destroyed entities.
When handling these operations, as triggered by commands, we assum the archetype-less entities have been reserved and never check it. This CL adds some checking so that we know when it happens in the future, but also the CL changes how we treat invalid entities while creating `FMassArchetypeEntityCollectionWithPayload`, which is the tool commands are using to batch together updates to diverse collections of entities.

This fix has been prompted by project-specific code issuing commands outside of Mass processing, and before Mass processing actually starts there's gameplay code that synchronously kills off relevant entities. That behavior is not recommended and the project-side code will be updated, but Mass needs to handle this case gracefully.

[CL 27833407 by mieszko zielinski in ue5-main branch]