PR #10839: Fix disappearing root motion in template sequences.
The montage data is mapped per MovieSceneSkeletalAnimationSection. The problem is that this object is the same if the template sequence is used for multipl objects, so the montage instance ID gets overwritten every frame. Thus, during the next frame when updating the anim positions, it cannot find the already existing montage that is playing on the anim instance, and creates a new one.

#rb ludovic.chabant
#jira UE-194772

[CL 27924889 by Victov in ue5-main branch]