Mobile forward light function bug fix and performance improvements
In the original changelist that added light function support to the mobile forward renderer the proxy geometry had face culling disabled in case the camera was outside the light function proxy geometry. This caused incorrect rendering (noticeable mainly when moving in and out of the light function geometry area) and also a performance penalty. This is fixed with this change.

This change also creates a stencil mask in the initial (non-lightfunction) pass to reduce the number of pixels shaded in the lightfunction draws.
This also enables depth test for lightfunction draws to further reduce the number of pixel shader executions.

#rb Florin.Pascu, Dmitriy.Dyomin

[CL 27958486 by florian penzkofer in ue5-main branch]