Fixed crash in paper2D due to DestroyRenderThreadResources not being called when a proxy is both added and removed from the scene prior to a scene update.
CreateRenderThreadResources is now deferred until the scene update when render command pipes are enabled in order to avoid race conditions when binding vertex factories. For symmetry, DestroyRenderThreadResources is only called if a corresponding CreateRenderThreadResources call will occur later during scene update. However, paper 2D initializes the vertex factories outside of the Create / Destroy pair and then expects Destroy to always be called. As it's not necessary to release render resources from this method anymore, they can safely be moved to the destructor and avoid the problem altogether.

[CL 27965447 by zach bethel in ue5-main branch]