Sequencer: start tracking pre-animated state on new track instance inputs before stopping tracking on old ones
When two track instance inputs correspond to two adjacent sections, we see the input of the ending section be removed, and the input of the starting section be added. This would potentially mean that the first section gets removed from pre-animated state tracking before the new section gets started with the same tracking. If the two sections happen to be tied to the same "thing", and are the only contributors, we would restore pre-animated state between the two for nothing. Now we correctly carry pre-animated state over to the second section.

#rb max.chen, andrew.rodham

[CL 27997884 by ludovic chabant in ue5-main branch]