Decommit mark bits when they aren't in use.
If the GC is running, then mark bits will definitely be committed. If it's not running, then the scavenger will eventually decommit them all. If the GC isn't running and you ask libpas to scavenge synchronously, then the mark bits will get decommitted.

This is designed so that in steady state, when GC runs regularly, the mark bits stay committed and nobody loops over them during GCs or scavenges. It's only if the GC goes idle that the scavenger will eventually loop over them, and of course the GC has to loop over them if we are re-committing.

#rb saam.barati

[CL 28022015 by phil pizlo in ue5-main branch]