[Json] Handle empty keys inside json objects.
Previously we would try to ignore empty string keys and serialize the values without it but that would trigger an assert inside json objects because only arrays can serialize values without keys. Empty strings are valid key strings in Json, so the previous implementation did not respect the specifications.

This change makes sure we only support empty keys when we're serializing a json object (arrays are excluded). Automated test cases have also been added to validate the empty key support and that we properly handle the case that was causing a crash during previous attempt at this fix (see backout CL 23028888).

#jira UE-195304
#rb Steve.Robb Johan.Torp

[CL 28033714 by benoit deschenes in ue5-main branch]