Sequencer: Defer deletion of sub sequence instances to the active runner
When a sequence is re-compiled mid-evaluation, we have to re-update the sequence updater to ensure that the set of imported entities is accurate. Previously, references to expiring sequence instances would be kept in the updater until the _next_ frame, when they would be removed from the map. Unfortunately this exposes a vulnerability where multiple calls to Update can occur, causing it to be destroyed prematurely on the second call (because bNeedsDestroy was made true on the first call), leaving entities still alive referencing that instance with NeedsUnlink until the end of the PostInstantiation.

The fix is to defer the destruction of the instance to the runner itself, like is true for root sequences. Unfortunately this does make FindRelatedInstanceHandle unable to work correctly for expiring sub sequences. In practice this is not really a problem because the only time we use that function is in the template sequence system, and that can reference things just as easily by RootInstanceHandle and SequenceID

#rb Max.Chen, Ludovic.Chabant

[CL 28118707 by andrew rodham in ue5-main branch]