#jira UE-193689 Removing a GameplayEffect after removing an AttributeSet that is affected by that GE resulted in a crash. This affects Standalone and servers when removing those two out of order. It also affects clients when SpawnedAttributes and ActiveGameplayEffectsContainer are replicated in that problematic order, despite that the server modified them in the correct order.
The crash occurs because after removing an attribute set, aggregators remain in the ActiveGameplayEffectsContainer for removed attributes. Calling OnDirty() on those aggregators on removing a GE then executes AbilitySystemComponent functions that assume that a matching attribute set exists.

With this change, when an attribute set is removed, aggregators for that set's attributes are removed from ActiveGameplayEffectsContainers immediately. Then, removing a gameplay effect afterwards will not attempt to access aggregators for attributes that don't actively have an attribute set in the ASC.

Also introduced a reusable FGameplayAttribute iteration by attribute set class, instead of copying the not-immediately-clear TFieldIterator approach that was used in several places.
#rb jodon.karlik

[CL 28130157 by zhikang shao in ue5-main branch]