Chaos - fix CollisionConstraint initial penetration MinInitialPhi
E.g., Objects would pop out of the ground even when InitialDepenetrationVelocity was set to zero. We don't save solver results for contacts that did not generate an impulse so we were throwing away initial penetration data. We must calculate and store MinimumInitialPhi on the previous frame when the data is still available.

#rb none

[CL 28278627 by chris caulfield in ue5-main branch]