Fix for a rare crash when a large number of objects are deleted in two consecutive frames.
At the root of the problem of the failure is a memory stomp in TEDS' scratch buffer. The scratch buffer assumes that all work has completed meaning no more of its memory is in use. Mass however queues commands during processing up for next frame so the scratch buffer releases memory too early in some case. The following criteria have to be met for the crash to trigger:
- Before the TEDS tick a large enough number of commands with extra data need to be created, at least enough to fill a scratch buffer chunk.
- At the start of the TEDS tick the filled memory chunks are recycled. This is the bug as the chunks may still be in use. This is by itself not enough to trigger the crash as the chunks are not cleared so the bug doesn't happen until the blocks are reused.
- During the TEDS tick new commands need to be issued before the TEDS tick reached a point where the chunk data is processed. This corrupts the original data to a degree that will trigger the crash.

The solution to the problem is to keep a frame counter in the scratch buffer and a last touched frame number in each chunk. During recycling only the chunks that were touched in the last frame or earlier can be recycled, while the rest will have to wait for another frame.

#jira UE-196780
#rb logan.buchy

[CL 28347553 by ronald koppers in ue5-main branch]