Fix hash collisions in raytracing autoinstanced mesh batches by including the shader hash
This was noticed on the substrate content sample, though in theory the issue could manifest in other situations as well. The hash used by autoinstancing only used the bindings and not the actual shader being used. So materials with identical paramters but different shading models or other settings could in theory result in collisions.

#jira UE-195868
#rb Aleksander.Netzel

[CL 28365719 by chris kulla in ue5-main branch]