Fix missing landscape in 120Hz mode.
When launching 120Hz, we swap between nanite on and nanite off mid way through the loading.

Issue was that Rendering is initialized in PostLoad and Proxy was registered in PostRegisterAllComponents (called by the streaming). On CVar change, we were iterating through the registered Proxies and calling UpdateRendering on them. But sometimes we could end up with Proxies initialized with Nanite On, but not yet registered when the CVar changed happened, leading to Proxy still using the nanite version.

Fix is to move the UpdateRendering alongside the register.

[REVIEW]

[CL 28413862 by john huelin in ue5-main branch]