Problem: Undo history fails to open after closing with it docked.
Root cause: We shouldn't be specifying saved positions for anything but positionally specified windows while  trying to invoke a tab. This code path was only put in to support repositioning specified windows from the layout, so any other type of EWindowPlacement is untested and unpredictable at this point.

Solution: Added a new method to TabManager which gets the proper FArea to load a tab. If it should be in a positionally specified window according to the initial layout, it is loaded there, otherwise it returns a new FArea to load it.



#jira UE-196434
#rb  Brooke.Hubert

[CL 28415397 by karen jirak in ue5-main branch]