Allow requests to be rewound and tried again
Note currently unused here. Request messages are built in
Activity.Buffer. Then the buffer was reset and used for the response
message and internal data. Now the internal data is in the activity and
response message is appended to buffer so the original request message
is can be resent if need be. A note about HangUp from a socket is
included albeit not implemented as yet.

#rb pl
#rnx
#jira

[CL 28451823 by martin ridgers in ue5-main branch]