Fixed assert when unparenting an entity.
TEDS favors simplicity so supports queueing up the deletion of columns that may not be on a row. This reduces complexity in the query callbacks as these do not need to first verify if a column is actually present. An example of this in action is the removal of FTypedElementParentColumn and FTypedElementUnresolvedParentColumn in one call when it's discovered that there's no longer a parent. Only one of these two columns will ever be active on a row.
Mass however takes a stricter approach and doesn't allow the removal of a fragment/tag from an archetype unless it's there. To bridge the two, TEDS will behind the scenes check for this case and filter out any fragments/tags that aren't on the archetype. To achieve this Mass had to get a small update to be able to access the required information.

#jira UE-197057
#rb brooke.hubert

[CL 28521283 by ronald koppers in ue5-main branch]