AnimNext: Workspace editor
Added a new editor that incorporates editing functionality for multiple assets at once. Assets are represented as multiple sub-document tabs across multiple tab wells. A CVar controls whether to open in the new workspace or in seperate sub-editors.
Reworked the AnimNextGraph a little to be more in-line with parameter blocks - we may be able to go further and merge them to a common base too.
Contains minor changes for RigVM to allow IRigVMClientHost to get corresponding RigVM graphs for editor graphs.

#rb Nicholas.Frechette, Jaime.Cifuentes, sara.schvartzman

[CL 28613272 by thomas sarkanen in ue5-main branch]