Export DRED breadcrumbs as part of the crash payload.
DRED breadcrumbs are processed into breadcrumb nodes first and then added to the crash payload, unless RHI breadcrumbs are enabled (in that case, RHI breadcrumbs are preferred).

A breadcrumb tree is reconstructed from the linearized DRED history. BeginEvent/EndEvent operations are used to mark the start/end of a scope (breadcrumb node with potential children). An event is marked as "Finished" if the EndEvent operation has been reported as completed (e.g. it's at or before the last completed operation as reported by DRED).

Note that especially when lightweight (markers only) DRED is active, the command history may begin and end at arbitrary points. This change should handle all possible cases, in particular the case where the history starts in the middle of a command queue where EndEvent operations do not have a matching BeginEvent operation before. In that case, we consider those events "Unknown" (since the contextual name is lost). For example, the following linear history where the final EndEvent does not have an associated BeginEvent:

```
1) BeginEvent [BeginOcclusionTests]
2) BeginEvent [ShadowFrustumQueries]
3) EndEvent
4) EndEvent
5) EndEvent
```

becomes

```
> Unknown Event
--- > BeginOcclusionTests
--- --- > ShadowFrustumQueries
```

#tests Tested that the produced breadcrumb tree from DRED matches the tree produced by RHI breadcrumbs (to the extent possible)
#rb mihnea.balta

[CL 28713303 by daniele vettorel in ue5-main branch]