Fixed undoing the removal of an object not causing to reregister that object with TEDS.
Undoing the deletion of an object will not always trigger that object to go through its full initialization path and as a result may not end up being registered with TEDS again. This change mirrors an earlier change that allowed correctly undoing a creation, however some more data was needed as the table isn't know at destruction. The table handle is now stored with the object in TEDS so it can be retrieved during removal and stored in an transaction command. This assumes that after restoring the object can be fully restored based on the information stored in the object. Additional information such as selection may need to be stored through the memento system, but this integration was left as future work on order to address the immediate bug.

#jira UE-197659
#rb semion.piskarev, logan.buchy

[CL 28727033 by ronald koppers in ue5-main branch]