Fix missing cards data on multi-group /multi-geometry type grooms.
For LOD0/LOD1, grooms have both strands & cards data. The LOD settings are set to strands geometry type for LOD0/LOD1 . When building the groom data, the cards data were discarded due to the mismatch of geometry type. It works for group0 as a matter of facts but all groups >0 would have not data being built.

#rb none

[CL 28752421 by charles derousiers in ue5-main branch]