Sequencer: resolve desired view target before checking if it's the same as the current one
When blending back to gameplay we pass a null pointer for the desired view target. We would later resolve this pointer with the pre-animated pointer, but not before checking if the desired view target was already set, in which case we would early-return. This caused the small issue that, once blending back to gameplay started, we should indeed early-return (since the camera change is already going on), but we weren't because we were checking against the null pointer, not against the resolved pre-animated pointer.

This change reorders a few things to make it all happen in the correct order.

#jira UE-197785
#rb max.chen

[CL 28777443 by ludovic chabant in ue5-main branch]