Cooker: GeneratedPackages: During garbage collection, for generator and generated packages, fix the condition for whether we need to keep all CachedObjectsInOuter referenced to be HasTakenOverCachedCookedPlatformData. The previous condition of GetSaveState() > CallObjectsToMove was incorrect and should have been >=. HasTakenOverCachedCookedPlatformData is the same thing as comparing with State >= CallObjectsToMove, but is more authoritative if we ever change the state where we do the take over. The incorrect comparison caused us not to keep the objects referenced, which allowed them to be GC'd, which breaks the CookPackageSplitter contract and also triggered an error in FGeneratorPackage::ResetSaveState when we detected that deletion and demoted the package out of save.
Cooker: GeneratedPackages: do not demote generator or generated packages when private objects are garbage collected. Demote  (and give an error) only  when public objects are garbage collected. Normal packages that are not a generator and are not owned by a generator keep the same behavior of demoting if any object, even a private one, is garbage collected.

Cooker: When a generated package is demoted due to deleted objects, add feedback with the name of the object that was deleted.

#jira UE-197955
#rbx

[CL 28932227 by matt peters in ue5-main branch]