Verse heap iteration should not hold onto libpas locks while calling callbacks.
Now, all iteration callbacks are called from the whatever thread does the iteration (i.e. the GC thread in VVM) and they are called with no locks held. The iteration algorithm carefully snapshots whatever it needs to know about the heap while holding some libpas locks and then does the actual iteration from that snapshot. Usually the snapshot is a 128-byte bitvector.

This change allows us to allocate while holding locks that would also be held in census or destruction.

It also allows census and destruction to call libpas API that might hold locks (like any libpas malloc, or possibly some functions that query the state of the verse heap).

This change does not mean you can allocate while iterating, though it does remove the main technical reason why we didn't allow that. Now we still don't allow it but only as a matter of policy.

#rb saam.barati

[CL 28935536 by phil pizlo in ue5-main branch]