- IAS could stop processing requests if it hit its fail timeout - New work was added in reverse order
IAS has a mechanism to fail requests if they're taking to long. This is to prevent live locks in very rare scenarios where network traffic comes to a complete halt. In the discovered repro case it involved using an external tool to silently drop traffic. This mechanism was incorrectly dropping an array such that there was nothing left to process requests still pending.

While fleshing out the tests I noticed that work that is added slightly out of order. This is mostly harmless due to the way IAS disaptches requests, but for best results in high-latency scenarios the mis-ordering was corrected by reversing a list.

Note that this change includes my test code changes. These are not compiled in Test or Shipping.

#rb pj
#rnx
#jira
#tests Paul Chipchase found a repro and confirmed the suggested fix resolves the issue. I improved my existing tests to cover this scenario and then implemented the fix. Win64/test locker, soak tested my library of replays overnight. Mega-compile across all platforms and variants.

[CL 28972815 by martin ridgers in ue5-main branch]